Splash Screen
Added a splash screen simply displaying the logo and launching the login
activity after 5 seconds. The logo can be tapped to immediately launch
the next activity. The handler callbacks are properly removed when the
app is closed to stop the next activity launching over other apps.

My plan is to use this activity to get stored user data and skip the
login if the credentials are still valid.