Sensor Preferences UI
Built the interface for the sensor preferences activity consisting of a
recycler view populated via the get all sensors endpoint. I included a
special click listener to allow me to toggle a nested checkbox when the
list item is tapped. Just need to wire the list up to the save endpoint