Vehicle Activity
I've built a starting UI for the vehicle activity including the usual
bits and pieces, and a custom recycler including warning and error
symbols. This will be linking into the actual graphs.

I also added in a custom Callback abstract class because I was writing
the same logic for every onError